# swagger/swagger-output.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "description": "Description",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:2222/"
        }
    ],
    "paths": {
...
        "/api/profile/": {
            "get": {
                "description": "",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "User profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "serverTime": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "profile": {
                                                    "$ref": "#/components/schemas/Profile"
                                                }
                                            },
                                            "required": ["profile"]
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "serverTime",
                                        "data"
                                    ],
                                    "xml": {
                                        "name": "main"
                                    }
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "serverTime": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "profile": {
                                                    "$ref": "#/components/schemas/Profile"
                                                }
                                            },
                                            "required": ["profile"]
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "serverTime",
                                        "data"
                                    ],
                                    "xml": {
                                        "name": "main"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "description": "",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Created user profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "serverTime": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "profile": {
                                                    "$ref": "#/components/schemas/Profile"
                                                }
                                            },
                                            "required": ["profile"]
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "serverTime",
                                        "data"
                                    ],
                                    "xml": {
                                        "name": "main"
                                    }
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "serverTime": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "profile": {
                                                    "$ref": "#/components/schemas/Profile"
                                                }
                                            },
                                            "required": ["profile"]
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "serverTime",
                                        "data"
                                    ],
                                    "xml": {
                                        "name": "main"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "firstName": {
                                        "example": "any"
                                    },
                                    "surname": {
                                        "example": "any"
                                    },
                                    "description": {
                                        "example": "any"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Updated user profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "serverTime": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "profile": {
                                                    "$ref": "#/components/schemas/Profile"
                                                }
                                            },
                                            "required": ["profile"]
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "serverTime",
                                        "data"
                                    ],
                                    "xml": {
                                        "name": "main"
                                    }
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "serverTime": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "profile": {
                                                    "$ref": "#/components/schemas/Profile"
                                                }
                                            },
                                            "required": ["profile"]
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "serverTime",
                                        "data"
                                    ],
                                    "xml": {
                                        "name": "main"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "firstName": {
                                        "example": "any"
                                    },
                                    "surname": {
                                        "example": "any"
                                    },
                                    "description": {
                                        "example": "any"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Deleted user profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "serverTime": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "deletedId": {
                                                    "type": "string",
                                                    "example": "1"
                                                }
                                            },
                                            "required": ["deletedId"]
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "serverTime",
                                        "data"
                                    ],
                                    "xml": {
                                        "name": "main"
                                    }
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "serverTime": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "deletedId": {
                                                    "type": "string",
                                                    "example": "1"
                                                }
                                            },
                                            "required": ["deletedId"]
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "serverTime",
                                        "data"
                                    ],
                                    "xml": {
                                        "name": "main"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/profile/{userId}": {
            "get": {
                "description": "",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "serverTime": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "profile": {
                                                    "$ref": "#/components/schemas/Profile"
                                                }
                                            },
                                            "required": ["profile"]
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "serverTime",
                                        "data"
                                    ],
                                    "xml": {
                                        "name": "main"
                                    }
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "serverTime": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "profile": {
                                                    "$ref": "#/components/schemas/Profile"
                                                }
                                            },
                                            "required": ["profile"]
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "serverTime",
                                        "data"
                                    ],
                                    "xml": {
                                        "name": "main"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Profile": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "1"
                    },
                    "firstName": {
                        "type": "string",
                        "example": "Kacper"
                    },
                    "surname": {
                        "type": "string",
                        "example": "Zakrzewski"
                    },
                    "description": {
                        "type": "string",
                        "example": "I'm a programmer"
                    },
                    "userId": {
                        "type": "string",
                        "example": "1"
                    },
                    "avatarUrl": {
                        "type": "string",
                        "example": "https://kzakrzewski.pl/avatar.png"
                    }
                },
                "required": [
                    "id",
                    "firstName",
                    "surname",
                    "description",
                    "userId"
                ],
                "xml": {
                    "name": "Profile"
                }
            },
...
        }
    }
}